refactor: 임시 매칭 정보 entity 명 변경
