--- src/main/java/com/dnlab/tacktogetherbackend/match/repository/TemporaryMatchInfoRepository.java
@@ -1,8 +1,0 @@
-package com.dnlab.tacktogetherbackend.match.repository;
-
-import com.dnlab.tacktogetherbackend.match.domain.TemporaryMatchInfo;
-import org.springframework.data.repository.CrudRepository;
-
-public interface TemporaryMatchInfoRepository extends CrudRepository<TemporaryMatchInfo, String> {
-    TemporaryMatchInfo deleteBySessionId(String sessionId);
-}
